requirements.txt
django
django-annoying
hiredis